refactor: Splitted AddNewMercahnt form
